Fixed non Owner can edit or delete image bug_version_3

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -91,14 +91,25 @@ public class ImageController {
 
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
+    //We Obtain Logged User info from session object using HttpSession
+    //We obtain the owner of the image using image id
+    //Later use if condition to Pop-up error as expected
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model,HttpSession session) {
         Image image = imageService.getImage(imageId);
-
-        String tags = convertTagsToString(image.getTags());
-        model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
-        return "images/edit";
+        User user =(User)session.getAttribute("loggeduser");
+        if(image.getUser().getId() != user.getId()){
+            String error = "Only the owner of the image can edit the image";
+            model.addAttribute("image",image);
+            model.addAttribute("tags",image.getTags());
+            model.addAttribute("editError",error);
+            return "images/image";
+        }else {
+            String tags = convertTagsToString(image.getTags());
+            model.addAttribute("image", image);
+            model.addAttribute("tags", tags);
+            return "images/edit";
+        }
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -132,17 +143,34 @@ public class ImageController {
         updatedImage.setDate(new Date());
 
         imageService.updateImage(updatedImage);
-        return "redirect:/images/" + updatedImage.getTitle();
+        return "redirect:/images/"+ updatedImage.getId()+"/"+updatedImage.getTitle();
     }
 
 
     //This controller method is called when the request pattern is of type 'deleteImage' and also the incoming request is of DELETE type
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
+    //We Obtain Logged User info from session object using HttpSession
+    //We obtain the owner of the image using image id
+    //Later use if condition to Pop-up error as expected
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId,Model model,HttpSession session) {
+
+        Image image = imageService.getImage(imageId);
+
+        User user = (User) session.getAttribute("loggeduser");
+
+        if(image.getUser().getId() != user.getId()){
+            String error = "Only the owner of the image can edit the image";
+
+            model.addAttribute("image", image);
+            model.addAttribute("tags", image.getTags());
+            model.addAttribute("deleteError", error);
+            return "images/image";
+        }else {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }
     }
